// src/meal/enums/meal-group-status.enum.ts
export enum MealGroupStatusEnum {
  CREATED = "created",
  MEALS_QUEUE = "meals_queue",
  MEALS_GENERATION = "meals_generation",
  MEAL_GENERATING = "meal_generating",
  READY_FOR_SHOPPING = "ready_for_shopping",
  AWAITING_SHOPPING_LIST = "awaiting_shopping_list",
  COMPLETE = "complete",
  FAILED = "failed",
}